Disable Crash Handler with Libgodot enabled.
The .net runtime relies on signals for its internal exception handling. In normal Godot operation, this is a non-issue because it gets loaded after Godot's signals. However, if you run Godot from C#, the signal handler is already in place once Godot starts and must not be disturbed. This could probably be cleaner.