Put GodotPlugins in the default load context.
Similar to the previous commit, any assembly linked to the main program will be dropped in the default load context, likely along with GodotSharp.dll. We want to avoid starting a second one in a different load context.

hostfxr's interfaces are questionably documented and horribly incomplete. As of this writing, load_assembly_and_get_function_pointer *always* puts the assembly in a custom isolated load context. load_assembly claims to take a parameter to specify load context, but in reality it doesn't support this, only verifying that the parameter is null and then putting it in the default context. Thankfully, this is the behavior we want anyway, so we use that to load the assembly and then get a function pointer out of it by hand.

(See [the HostFXR source code](https://github.com/dotnet/runtime/blob/0edefd061f456fa6021b36e341b392bbd6eff85c/src/libraries/System.Private.CoreLib/src/Internal/Runtime/InteropServices/ComponentActivator.cs#L113) for validation.)